+++ src/tsp/tsp.h
@@ -23,7 +23,7 @@ class TSP : public QGraphicsScene, public GAOptimizationProblem<int>
     std::vector<Coordinate> getTargets();
     void setPath(const std::vector<Coordinate> &steps);
     unsigned int getSolutionSize() const;
-    double evaluateSolution(const std::vector<int> &s) const;
+    long double evaluateSolution(const std::vector<int> &s) const;
     void showSolution(const std::vector<int> &s);
     std::vector<int> crossover(const std::vector<int> &a, const std::vector<int> &b);
     void mutate(std::vector<int> &);
